cloud_function/requirements.txt
functions-framework==3.*
google-cloud-firestore
firebase-admin
requests
google-auth
google-cloud-aiplatform
flask